+++ src/components/filterform/index.ts
@@ -1,3 +1,2 @@
 export { FiltersForm } from './FiltersForm'
-export type { FilterForm } from './FiltersForm'
-export type { FilterConfig, FilterType } from './FilterField'
+export type { FilterForm, FilterConfig, FilterType, Filter } from './types'
